backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
sqlalchemy
google-generativeai
python-dotenv
